+++ РПГ/Player.cpp
@@ -1,3 +1,4 @@
+#pragma once
 #include "Player.h"
 using namespace std;
 Player create()
